+++ packData/packData.mjs
@@ -25,9 +25,12 @@ let packs = [
     'cpr-summon-features-2024',
     'cpr-feats-2024',
     //'cpr-feat-features-2024',
+    'cpr-feat-features-2024',
     'cpr-embedded-macro-sample-items',
     'cpr-actions-2024',
-    'cpr-feature-items-2024'
+    'cpr-feature-items-2024',
+    'cpr-3rd-party-class-features-2024',
+    'cpr-3rd-party-spells'
 ];
 for (let i of packs) {
     await compilePack('./packData/' + i, './packs/' + i, {'log': true});