+++ scripts/macros/2014/classFeatures/bloodHunter/brandOfCastigation.js
@@ -1,5 +1,4 @@
 import {activityUtils, actorUtils, compendiumUtils, constants, dialogUtils, effectUtils, errors, genericUtils, itemUtils, tokenUtils, workflowUtils} from '../../../../utils.js';
-
 async function late({trigger: {entity: item}, workflow}) {
     if (!workflow.hitTargets.size) return;
     if (workflow.item.type !== 'weapon') return;
@@ -26,7 +25,7 @@ async function late({trigger: {entity: item}, workflow}) {
     };
     let feature = activityUtils.getActivityByIdentifier(item, 'brandOfCastigation', {strict: true});
     if (!feature) return;
-    await workflowUtils.syntheticActivityRoll(feature, [], {config: {consumeUsage: true}});
+    await workflowUtils.syntheticActivityRoll(feature, [], {consumeResources: true, consumeUsage: true});
     effect = await effectUtils.createEffect(workflow.actor, effectData, {identifier: 'brandOfCastigationSource'});
     if (!effect) return;
     effectData.flags = {
@@ -38,6 +37,7 @@ async function late({trigger: {entity: item}, workflow}) {
     await effectUtils.createEffect(workflow.targets.first().actor, effectData, {identifier: 'brandOfCastigation', parentEntity: effect, strictlyInterdependent: true});
 }
 async function hit({trigger: {entity: effect}, workflow}) {
+    if (!workflow.damageRolls.length || !workflow.hitTargets.size) return;
     let effects = effectUtils.getAllEffectsByIdentifier(workflow.actor, 'brandOfCastigation');
     if (!effects.length) return;
     if (!effects.filter(i => i.origin === effect.origin).length) return;
@@ -48,6 +48,7 @@ async function hit({trigger: {entity: effect}, workflow}) {
     await workflowUtils.syntheticActivityRoll(feature, [workflow.token]);
 }
 async function hitNear({trigger: {entity: effect}, workflow}) {
+    if (!workflow.damageRolls.length || !workflow.hitTargets.size) return;
     let effects = effectUtils.getAllEffectsByIdentifier(workflow.actor, 'brandOfCastigation');
     if (!effects.length) return;
     if (!effects.filter(i => i.origin === effect.origin).length) return;