+++ scripts/macros/2014/classFeatures/paladin/oathOfDevotion/sacredWeapon.js
@@ -71,6 +71,7 @@ async function use({workflow}) {
 }
 export let sacredWeapon = {
     name: 'Channel Divinity: Sacred Weapon',
+    aliases: ['Sacred Weapon'],
     version: '1.1.0',
     midi: {
         item: [