+++ scripts/macros/2014/spells/absorbElements.js
@@ -1,0 +1,5 @@
+export let absorbElements = {
+    name: 'Absorb Elements',
+    version: '1.3.105',
+    rules: 'legacy'
+};