+++ scripts/macros/2014/spells/veilOfDusk.js
@@ -1,0 +1,5 @@
+export let veilOfDusk = {
+    name: 'Veil of Dusk',
+    version: '1.3.105',
+    rules: 'legacy'
+};