+++ scripts/macros/2024/classFeatures/warlock/magicalCunning.js
@@ -1,0 +1,5 @@
+export let magicalCunning = {
+    name: 'Magical Cunning',
+    version: '1.3.105',
+    rules: 'modern',
+};